+++ ComponentApi/src/main/java/com/xiaojinzi/component/anno/InterceptorAnno.java
@@ -1,6 +1,6 @@
 package com.xiaojinzi.component.anno;
 
-import com.xiaojinzi.component.anno.support.MainThreadCreateAnno;
+import com.xiaojinzi.component.anno.support.UiThreadCreateAnno;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -10,7 +10,7 @@ import java.lang.annotation.Target;
 /**
  * 拦截器的注解,用这个注解标记的不是全局的拦截器,但是你可以使用特定的字符串访问到这个拦截器
  */
-@MainThreadCreateAnno
+@UiThreadCreateAnno
 @Target({ElementType.TYPE})
 @Retention(RetentionPolicy.CLASS)
 public @interface InterceptorAnno {