+++ ComponentApi/src/main/java/com/xiaojinzi/component/anno/ServiceParameterAnno.java
@@ -1,6 +1,6 @@
 package com.xiaojinzi.component.anno;
 
-import com.xiaojinzi.component.anno.support.MainThreadCreateAnno;
+import com.xiaojinzi.component.anno.support.UiThreadCreateAnno;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -10,7 +10,7 @@ import java.lang.annotation.Target;
 /**
  * 标记一个 Service 服务接口的方法的参数是暴露的
  */
-@MainThreadCreateAnno
+@UiThreadCreateAnno
 @Target({ElementType.TYPE, ElementType.PARAMETER})
 @Retention(RetentionPolicy.CLASS)
 public @interface ServiceParameterAnno {