+++ ComponentImpl/src/main/java/com/xiaojinzi/component/impl/RouterCenter.java
@@ -5,9 +5,9 @@ import android.app.Application;
 import android.content.Context;
 import android.content.Intent;
 import android.net.Uri;
-import android.support.annotation.MainThread;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
+import android.support.annotation.UiThread;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentActivity;
 import android.support.v4.app.FragmentManager;
@@ -16,6 +16,7 @@ import com.xiaojinzi.component.Component;
 import com.xiaojinzi.component.ComponentConstants;
 import com.xiaojinzi.component.ComponentUtil;
 import com.xiaojinzi.component.anno.support.CheckClassNameAnno;
+import com.xiaojinzi.component.anno.support.NotAppUseAnno;
 import com.xiaojinzi.component.bean.PageInterceptorBean;
 import com.xiaojinzi.component.bean.RouterBean;
 import com.xiaojinzi.component.error.ignore.InterceptorNotFoundException;
@@ -53,6 +54,7 @@ import static com.xiaojinzi.component.ComponentConstants.SEPARATOR;
  * @author xiaojinzi
  * @hide
  */
+@NotAppUseAnno
 @CheckClassNameAnno
 public class RouterCenter implements IComponentCenterRouter {
 
@@ -96,7 +98,7 @@ public class RouterCenter implements IComponentCenterRouter {
     }
 
     @Override
-    @MainThread
+    @UiThread
     public void openUri(@NonNull final RouterRequest routerRequest) throws Exception {
         doOpenUri(routerRequest);
     }
@@ -105,7 +107,7 @@ public class RouterCenter implements IComponentCenterRouter {
      * @param request             路由请求对象
      * @param routerDegradeIntent 一个降级的 Intent
      */
-    @MainThread
+    @UiThread
     public void routerDegrade(@NonNull RouterRequest request, @NonNull Intent routerDegradeIntent) throws Exception {
         String uriString = request.uri.toString();
         if (routerDegradeIntent == null) {
@@ -119,7 +121,7 @@ public class RouterCenter implements IComponentCenterRouter {
      *
      * @param request 路由请求对象
      */
-    @MainThread
+    @UiThread
     private void doOpenUri(@NonNull final RouterRequest request) throws Exception {
         if (!Utils.isMainThread()) {
             throw new NavigationFailException("Router must run on main thread");
@@ -142,7 +144,8 @@ public class RouterCenter implements IComponentCenterRouter {
         Context rawContext = request.getRawContext();
         // 如果 Context 和 Fragment 中的 Context 都是 null
         if (rawContext == null) {
-            throw new NavigationFailException("is your fragment or Activity is Destoried?");
+            throw new NavigationFailException("is your fragment or Activity is Destoried?\n" +
+                    "see " + Component.ROUTER_UES_NOTE);
         }
         Intent intent = null;
         if (target.getTargetClass() != null) {
@@ -162,7 +165,7 @@ public class RouterCenter implements IComponentCenterRouter {
      * @param request 请求对象
      * @param intent  Intent
      */
-    @MainThread
+    @UiThread
     private void doStartIntent(@NonNull RouterRequest request,
                                Intent intent) throws Exception {
         // 前置工作
@@ -226,7 +229,7 @@ public class RouterCenter implements IComponentCenterRouter {
                 } else {
                     throw new NavigationFailException("the context or fragment both are null");
                 }
-            }else { // startActivityForResult
+            } else { // startActivityForResult
                 Activity rawAct = null;
                 if ((rawAct = Utils.getActivityFromContext(request.context)) != null) {
                     rawAct.startActivityForResult(intent, request.requestCode, request.options);
@@ -340,7 +343,7 @@ public class RouterCenter implements IComponentCenterRouter {
      * 找到那个 Activity 中隐藏的一个 Fragment,如果找的到就会用这个 Fragment 拿来跳转
      */
     @Nullable
-    private Fragment findFragment(Context context) {
+    private Fragment findFragment(@NonNull Context context) {
         Fragment result = null;
         Activity act = Utils.getActivityFromContext(context);
         if (act instanceof FragmentActivity) {
@@ -351,36 +354,35 @@ public class RouterCenter implements IComponentCenterRouter {
     }
 
     @Nullable
-    private Fragment findFragment(Fragment fragment) {
-        Fragment result = null;
-        if (fragment != null) {
-            result = fragment.getChildFragmentManager().findFragmentByTag(ComponentConstants.ACTIVITY_RESULT_FRAGMENT_TAG);
-        }
-        return result;
+    private Fragment findFragment(@NonNull Fragment fragment) {
+        return fragment.getChildFragmentManager().findFragmentByTag(ComponentConstants.ACTIVITY_RESULT_FRAGMENT_TAG);
     }
 
     @Override
-    public void register(IComponentHostRouter router) {
-        if (router == null) {
-            return;
+    public void register(@NonNull IComponentHostRouter router) {
+        Utils.checkNullPointer(router);
+        if (!hostRouterMap.containsKey(router.getHost())) {
+            hostRouterMap.put(router.getHost(), router);
+            routerMap.putAll(router.getRouterMap());
         }
-        hostRouterMap.put(router.getHost(), router);
-        routerMap.putAll(router.getRouterMap());
     }
 
     @Override
     public void register(@NonNull String host) {
-        IComponentHostRouter router = findUiRouter(host);
-        register(router);
+        Utils.checkStringNullPointer(host, "host");
+        if (!hostRouterMap.containsKey(host)) {
+            IComponentHostRouter router = findUiRouter(host);
+            if (router != null) {
+                register(router);
+            }
+        }
     }
 
     @Override
-    public void unregister(IComponentHostRouter router) {
-        if (router == null) {
-            return;
-        }
-        hostRouterMap.remove(router.getHost());
-        Map<String, RouterBean> childRouterMap = router.getRouterMap();
+    public void unregister(@NonNull IComponentHostRouter hostRouter) {
+        Utils.checkNullPointer(hostRouter);
+        hostRouterMap.remove(hostRouter);
+        Map<String, RouterBean> childRouterMap = hostRouter.getRouterMap();
         if (childRouterMap != null) {
             // key = host/path
             for (String key : childRouterMap.keySet()) {
@@ -391,8 +393,11 @@ public class RouterCenter implements IComponentCenterRouter {
 
     @Override
     public void unregister(@NonNull String host) {
-        IComponentHostRouter router = hostRouterMap.remove(host);
-        unregister(router);
+        Utils.checkStringNullPointer(host, "host");
+        IComponentHostRouter hostRouter = hostRouterMap.get(host);
+        if (hostRouter != null) {
+            unregister(hostRouter);
+        }
     }
 
     /**
@@ -422,7 +427,7 @@ public class RouterCenter implements IComponentCenterRouter {
         Set<String> set = new HashSet<>();
         for (Map.Entry<String, IComponentHostRouter> entry : hostRouterMap.entrySet()) {
             IComponentHostRouter childRouter = entry.getValue();
-            if (childRouter == null || childRouter.getRouterMap() == null) {
+            if (childRouter == null) {
                 continue;
             }
             Map<String, RouterBean> childRouterMap = childRouter.getRouterMap();