+++ ComponentImpl/src/main/java/com/xiaojinzi/component/impl/application/ModuleManager.java
@@ -2,8 +2,10 @@ package com.xiaojinzi.component.impl.application;
 
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
+import android.support.annotation.UiThread;
 
 import com.xiaojinzi.component.Component;
+import com.xiaojinzi.component.Config;
 import com.xiaojinzi.component.ComponentUtil;
 import com.xiaojinzi.component.application.IComponentCenterApplication;
 import com.xiaojinzi.component.application.IComponentHostApplication;
@@ -55,32 +57,34 @@ public class ModuleManager implements IComponentCenterApplication {
 
     @Override
     public void register(@NonNull IComponentHostApplication moduleApp) {
+        Utils.checkNullPointer(moduleApp);
         if (moduleApplicationMap.containsKey(moduleApp.getHost())) {
             LogUtil.loge("The module \"" + moduleApp.getHost() + "\" is already registered");
+        } else {
+            moduleApplicationMap.put(moduleApp.getHost(), moduleApp);
+            moduleApp.onCreate(Component.getApplication());
+            notifyModuleChanged();
         }
-        moduleApplicationMap.put(moduleApp.getHost(), moduleApp);
-        moduleApp.onCreate(Component.getApplication());
     }
 
     @Override
     public void register(@NonNull String host) {
-        if (host == null || "".equals(host)) {
-            throw new RuntimeException("the host can't be null or empty");
-        }
+        Utils.checkNullPointer(host, "host");
         if (moduleApplicationMap.containsKey(host)) {
-            LogUtil.log("the host '" + host + "' is already load");
+            LogUtil.loge("the host '" + host + "' is already load");
             return;
-        }
-        IComponentHostApplication moduleApplication = findModuleApplication(host);
-        if (moduleApplication == null) {
-            LogUtil.log("模块 '" + host + "' 加载失败");
         } else {
-            register(moduleApplication);
+            IComponentHostApplication moduleApplication = findModuleApplication(host);
+            if (moduleApplication == null) {
+                LogUtil.log("模块 '" + host + "' 加载失败, 请根据链接中的内容自行排查! \n " + Component.COMMON_ERROR_ISSUE);
+            } else {
+                register(moduleApplication);
+            }
         }
     }
 
     /**
-     * 自动注册, 需要开启 {@link com.xiaojinzi.component.Config.Builder#optimizeInit(boolean)}
+     * 自动注册, 需要开启 {@link Config.Builder#optimizeInit(boolean)}
      * 表示使用 Gradle 插件优化初始化
      */
     public void autoRegister() {
@@ -101,23 +105,17 @@ public class ModuleManager implements IComponentCenterApplication {
     public void registerArr(@Nullable String... hosts) {
         if (hosts != null) {
             for (String host : hosts) {
-                IComponentHostApplication moduleApplication = findModuleApplication(host);
-                if (moduleApplication == null) {
-                    LogUtil.log("模块 '" + host + "' 加载失败");
-                } else {
-                    register(moduleApplication);
-                }
+                register(host);
             }
         }
     }
 
     @Override
-    public void unregister(@Nullable IComponentHostApplication moduleApp) {
-        if (moduleApp == null) {
-            return;
-        }
+    public void unregister(@NonNull IComponentHostApplication moduleApp) {
+        Utils.checkNullPointer(moduleApp);
         moduleApplicationMap.remove(moduleApp.getHost());
         moduleApp.onDestroy();
+        notifyModuleChanged();
     }
 
     @Override
@@ -166,6 +164,30 @@ public class ModuleManager implements IComponentCenterApplication {
             }
         }
         return result;
+    }
+
+    @UiThread
+    private void notifyModuleChanged() {
+        final int compareValue = Utils.COUNTER.incrementAndGet();
+        Utils.postDelayActionToMainThread(new Runnable() {
+            @Override
+            public void run() {
+                // 说明没有改变过
+                if (compareValue == Utils.COUNTER.get()) {
+                    // LogUtil.loge("通知 " + compareValue);
+                    doNotifyModuleChanged();
+                } else {
+                    // LogUtil.loge("放弃通知 " + compareValue);
+                }
+            }
+        }, Component.getConfig().getNotifyModuleChangedDelayTime());
+    }
+
+    @UiThread
+    private void doNotifyModuleChanged() {
+        for (IComponentHostApplication hostApplication : moduleApplicationMap.values()) {
+            hostApplication.onModuleChanged(Component.getApplication());
+        }
     }
 
     /**