+++ ComponentImpl/src/main/java/com/xiaojinzi/component/support/IBaseLifecycle.java
@@ -1,22 +1,22 @@
 package com.xiaojinzi.component.support;
 
 import android.app.Application;
-import android.support.annotation.MainThread;
 import android.support.annotation.NonNull;
+import android.support.annotation.UiThread;
 
 import com.xiaojinzi.component.anno.support.CheckClassNameAnno;
 
 /**
  * 基础的生命周期接口
  */
-@MainThread
+@UiThread
 @CheckClassNameAnno
 public interface IBaseLifecycle {
 
-    @MainThread
+    @UiThread
     void onCreate(@NonNull Application app);
 
-    @MainThread
+    @UiThread
     void onDestroy();
 
 }