+++ app/src/main/java/com/xiaojinzi/componentdemo/App.java
@@ -2,7 +2,6 @@ package com.xiaojinzi.componentdemo;
 
 import android.app.Application;
 
-import com.google.gson.Gson;
 import com.xiaojinzi.component.Component;
 import com.xiaojinzi.component.Config;
 import com.xiaojinzi.component.impl.application.ModuleManager;
@@ -27,12 +26,13 @@ public class App extends Application {
                         .routeRepeatCheckDuration(1000)
                         // 是否打印日志提醒你哪些路由使用了 Application 为 Context 进行跳转
                         .tipWhenUseApplication(true)
+                        .notifyModuleChangedDelayTime(200)
                         // 开启启动优化, 必须配套使用 Gradle 插件
                         .optimizeInit(true)
                         // 自动加载所有模块, 依赖上面的 optimizeInit(true)
                         .autoRegisterModule(true)
                         // demo 测试, 线上并没有, 请勿配置
-                        .objectToJsonConverter(obj -> new Gson().toJson(obj))
+                        // .objectToJsonConverter(obj -> new Gson().toJson(obj))
                         // 执行构建
                         .build()
         );